requirements.txt
datasets==3.0.2
numpy==2.3.3
PyYAML>=6.0
sentencepiece==0.2.0
torch==2.1.2
tqdm==4.66.4
transformers==4.46.1
wandb==0.22.0
